requirements.txt
Django==4.1.7
django-bootstrap-v5==1.0.11
django-extensions==3.2.1
django-extra-views==0.14.0
django-filter==22.1
django-seed==0.3.1
Pillow==9.4.0
whitenoise==6.3.0